+++ packages/forms/test/components/Form/RadioButtonController.test.tsx
@@ -1,39 +1,40 @@
 import React from 'react';
-import { shallow } from 'enzyme';
+import { mount } from 'enzyme';
 import BaseRadioButtonController from '@airbnb/lunar/lib/components/RadioButtonController';
-import { unwrapHOCs } from '@airbnb/lunar-test-utils';
 import RadioButtonController from '../../../src/components/Form/RadioButtonController';
 import { toString } from '../../../src/helpers';
+import { Context } from '../../../src/types';
+import { WrappingFormComponent, createFormContext } from '../../utils';
 
 describe('<RadioButtonController />', () => {
-  const form = {
-    change() {},
-    getState: () => ({} as any),
-    register: jest.fn(),
-  };
+  let context: Context;
+
+  beforeEach(() => {
+    context = createFormContext();
+  });
 
   it('connects to the form', () => {
-    const wrapper = unwrapHOCs(
-      shallow(
-        <RadioButtonController label="Label" name="foo" defaultValue="bar" validator={() => {}}>
-          {RB => (
-            <div>
-              <RB value="foo" label="Foo" />
-              <RB value="bar" label="Bar" />
-              <RB value="baz" label="Baz" />
-            </div>
-          )}
-        </RadioButtonController>,
-      ),
-      'FormRadioButtonController',
-      form,
+    const wrapper = mount(
+      <RadioButtonController label="Label" name="foo" defaultValue="bar" validator={() => {}}>
+        {RB => (
+          <div>
+            <RB value="foo" label="Foo" />
+            <RB value="bar" label="Bar" />
+            <RB value="baz" label="Baz" />
+          </div>
+        )}
+      </RadioButtonController>,
+      {
+        wrappingComponent: WrappingFormComponent,
+        wrappingComponentProps: { context },
+      },
     );
 
-    expect(form.register).toHaveBeenCalledWith(
+    expect(context.register).toHaveBeenCalledWith(
       expect.objectContaining({ name: 'foo', defaultValue: 'bar', parse: toString }),
       expect.anything(),
     );
 
-    expect(wrapper.shallow().find(BaseRadioButtonController)).toHaveLength(1);
+    expect(wrapper.find(BaseRadioButtonController)).toHaveLength(1);
   });
 });